make a class for the analytic pitch distributions
