+++ include/DREAM/Settings/OptionConstants.enum.hpp
@@ -148,6 +148,7 @@ enum collqty_pstar_mode {                // Runaway growth rates are determined 
 };
 
 enum collqty_Eceff_mode {
+    COLLQTY_ECEFF_MODE_EC_TOT = 0,      // Gives Ectot including all bound electrons (or Ec_free if no impurities/complete screening)
     COLLQTY_ECEFF_MODE_CYLINDRICAL = 1, // Sets Eceff using the Hesslow formula ignoring trapping effects.
     COLLQTY_ECEFF_MODE_SIMPLE = 2,      // An approximate numerical calculation with a simplified account of trapping effects
     COLLQTY_ECEFF_MODE_FULL = 3         // Full 'Lehtinen theory' expression.
