+++ src/CMakeLists.txt
@@ -24,17 +24,14 @@ set(dream_core
     "${PROJECT_SOURCE_DIR}/src/UnknownQuantityEquation.cpp"
 )
 set(dream_equations
+    "${PROJECT_SOURCE_DIR}/src/Equations/AnalyticDistribution.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/CollisionQuantityHandler.cpp"
-    "${PROJECT_SOURCE_DIR}/src/Equations/ConnorHastie.cpp"
-    "${PROJECT_SOURCE_DIR}/src/Equations/DreicerNeuralNetwork.cpp"
-    "${PROJECT_SOURCE_DIR}/src/Equations/RunawayFluid.cpp"
-    "${PROJECT_SOURCE_DIR}/src/Equations/EffectiveCriticalField.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/CollisionQuantity.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/CollisionFrequency.cpp"
-    "${PROJECT_SOURCE_DIR}/src/Equations/SlowingDownFrequency.cpp"
-    "${PROJECT_SOURCE_DIR}/src/Equations/PitchScatterFrequency.cpp"
-    "${PROJECT_SOURCE_DIR}/src/Equations/ParallelDiffusionFrequency.cpp"
+    "${PROJECT_SOURCE_DIR}/src/Equations/ConnorHastie.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/CoulombLogarithm.cpp"
+    "${PROJECT_SOURCE_DIR}/src/Equations/DreicerNeuralNetwork.cpp"
+    "${PROJECT_SOURCE_DIR}/src/Equations/EffectiveCriticalField.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/FluidSourceTerm.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/Fluid/DensityFromDistributionFunction.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/Fluid/CurrentDensityFromDistributionFunction.cpp"
@@ -60,9 +57,13 @@ set(dream_equations
     "${PROJECT_SOURCE_DIR}/src/Equations/Kinetic/AvalancheSourceRP.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/Kinetic/SlowingDownTerm.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/Kinetic/SynchrotronTerm.cpp"
+    "${PROJECT_SOURCE_DIR}/src/Equations/ParallelDiffusionFrequency.cpp"
+    "${PROJECT_SOURCE_DIR}/src/Equations/PitchScatterFrequency.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/PoloidalFlux/AmperesLawDiffusionTerm.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/PoloidalFlux/AmperesLawBoundaryAtRMax.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/PoloidalFlux/HyperresistiveDiffusionTerm.cpp"
+    "${PROJECT_SOURCE_DIR}/src/Equations/SlowingDownFrequency.cpp"
+    "${PROJECT_SOURCE_DIR}/src/Equations/RunawayFluid.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/TransportPrescribed.cpp"
     "${PROJECT_SOURCE_DIR}/src/Equations/TransportPrescribedBC.cpp"
 )
