+++ src/Equations/EffectiveCriticalField.cpp
@@ -1,10 +1,20 @@
 /**
  * Calculates the effective critical electric field
+ * Settings parameters: 
+ * - collQtySettings->collfreq_type (completely screened or not)
+ * - Eceff_mode: 
+ *  COLLQTY_ECEFF_MODE_EC_TOT = 0,      // Gives Ectot including all bound electrons (or Ec_free if no impurities/complete screening)
+ *  COLLQTY_ECEFF_MODE_CYLINDRICAL = 1, // Sets Eceff using the Hesslow formula ignoring trapping effects.
+ *  COLLQTY_ECEFF_MODE_SIMPLE = 2,      // An approximate numerical calculation with a simplified account of trapping effects
+ *  COLLQTY_ECEFF_MODE_FULL = 3         // Full 'Lehtinen theory' expression.
+ * 
+ * 
  */
-
-#include "DREAM/Equations/RunawayFluid.hpp"
 #include "FVM/config.h"
+#include "DREAM/Equations/AnalyticDistribution.hpp"
+#include "DREAM/Equations/RunawayFluid.hpp"
 #include "DREAM/Equations/EffectiveCriticalField.hpp"
+
 
 using namespace DREAM;
 
@@ -12,13 +22,10 @@ using namespace DREAM;
  * Constructor.
  *
  */
-EffectiveCriticalField::EffectiveCriticalField(Param1 *par,Param2 *par2) // @@@Linnea move collSettingsForEc?
-    : nr(par->rGrid->GetNr()), 
-    collQtySettings(par2->collQtySettings),
-    ions(par2->ions), fsolve(par2->fsolve), Eceff_mode(par2->Eceff_mode),
+EffectiveCriticalField::EffectiveCriticalField(Param1 *par,Param2 *par2)
+    : collQtySettings(par2->collQtySettings), fsolve(par2->fsolve), Eceff_mode(par2->Eceff_mode),
     rGrid(par->rGrid), nuS(par->nuS), nuD(par->nuD),
     fgType(par->fgType), gsl_ad_w(par->gsl_ad_w),fmin(par->fmin),collSettingsForEc(par->collSettingsForEc){}
-
 
 /**
  * Calculates and stores the effective critical field for runaway generation.
@@ -30,11 +37,13 @@ EffectiveCriticalField::EffectiveCriticalField(Param1 *par,Param2 *par2) // @@@L
  * is the net momentum advection term averaged over an analytic pitch
  * angle distribution.
  */
-void EffectiveCriticalField::CalculateEffectiveCriticalField(const real_t *Ec_tot, const real_t *Ec_free, real_t *effectiveCriticalField){ // @@Linnea real_t effectiveCriticalField[]?
-    if(Eceff_mode == OptionConstants::COLLQTY_ECEFF_MODE_CYLINDRICAL){
+void EffectiveCriticalField::CalculateEffectiveCriticalField(const real_t *Ec_tot, const real_t *Ec_free, real_t *effectiveCriticalField){
+    len_t nr = rGrid->GetNr();
+    // why if-else-return here and not below (l.42 and onwards?) I think it makes more sense to have two returns or one big if statement 
+    if(Eceff_mode == OptionConstants::COLLQTY_ECEFF_MODE_CYLINDRICAL){ 
         if(collQtySettings->collfreq_type==OptionConstants::COLLQTY_COLLISION_FREQUENCY_TYPE_COMPLETELY_SCREENED)
             for(len_t ir=0; ir<nr; ir++)
-                effectiveCriticalField[ir] = Ec_free[ir];
+                effectiveCriticalField[ir] = Ec_free[ir]; // should we account for radiation also if completely screened?
         else
             for(len_t ir=0; ir<nr; ir++)
                 effectiveCriticalField[ir] = Ec_tot[ir];
@@ -46,9 +55,10 @@ void EffectiveCriticalField::CalculateEffectiveCriticalField(const real_t *Ec_to
 
     real_t ELo, EUp;
     gsl_function UExtremumFunc;
-    for (len_t ir=0; ir<this->nr; ir++){
+    AnalyticDistribution analyticDist = AnalyticDistribution(rGrid, nuD, Eceff_mode);
+    for (len_t ir=0; ir<nr; ir++){
         UContributionParams params = {rGrid, nuS,nuD, ir, p, FVM::FLUXGRIDTYPE_DISTRIBUTION, Eterm, Func, gsl_ad_w,
-                            fmin, p_ex_lo, p_ex_up,collSettingsForEc,QAG_KEY,this};
+                            fmin, p_ex_lo, p_ex_up,collSettingsForEc,QAG_KEY,&analyticDist};
         UExtremumFunc.function = &(FindUExtremumAtE);
         UExtremumFunc.params = &params;
 
@@ -138,9 +148,6 @@ void EffectiveCriticalField::FindPExInterval(real_t *p_ex_guess, real_t *p_ex_lo
         }
 }
 
-//#include "RunawayFluid.UFunc.cpp"
-/////////
-
 /**
  * U(r,p) is the pitch-angle averaged momentum-advection coefficient,
  * U = < {A^p}f > / <f>,
@@ -148,74 +155,6 @@ void EffectiveCriticalField::FindPExInterval(real_t *p_ex_guess, real_t *p_ex_lo
  * The calculation is documented under doc/notes/theory.pdf in 
  * Section 2 (under the heading 'Bounce-averaged effective field') 
  */
-
-/**
- * Returns xi0/<xi> (the integral of which appears in AnalyticPitchDistribution).
- */
-struct distExponentParams {len_t ir; FVM::RadialGrid *rGrid;};
-real_t distExponentIntegral(real_t xi0, void *par){
-    struct distExponentParams *params = (struct distExponentParams *) par;
-    std::function<real_t(real_t,real_t,real_t)> xiFunc = [xi0](real_t BOverBmin, real_t , real_t )
-                            {return sqrt(1 - BOverBmin*(1-xi0*xi0));};
-    len_t ir = params->ir;
-    FVM::RadialGrid *rGrid = params->rGrid;
-    real_t signXi0 = ( (xi0>0) - (xi0<0));
-    real_t xiAvg = signXi0 * rGrid->CalculateFluxSurfaceAverage(ir,FVM::FLUXGRIDTYPE_DISTRIBUTION, xiFunc);
-    return xi0/xiAvg;
-}
-
-
-/**
- * Calculates the (semi-)analytic pitch-angle distribution predicted in the 
- * near-threshold regime, where the momentum flux is small compared 
- * to the characteristic pitch flux, and we obtain the approximate 
- * kinetic equation phi_xi = 0.
- */
-real_t EffectiveCriticalField::evaluateAnalyticPitchDistribution(len_t ir, real_t xi0, real_t p, real_t Eterm, CollisionQuantity::collqty_settings *inSettings, gsl_integration_workspace *gsl_ad_w){
-    const real_t Bmin = rGrid->GetBmin(ir);
-    const real_t Bmax = rGrid->GetBmax(ir);
-    const real_t B2avgOverBmin2 = rGrid->GetFSA_B2(ir);
-    real_t xiT = sqrt(1-Bmin/Bmax);
-    real_t E = Constants::ec * Eterm / (Constants::me * Constants::c) * sqrt(B2avgOverBmin2); 
-
-    real_t pNuD = p*nuD->evaluateAtP(ir,p,inSettings);    
-    real_t A = 2*E/pNuD;
-
-    // This block carries defines the integration int(xi0/<xi(xi0)> dxi0, xi1, x2) 
-    //////////////////////////////
-    gsl_function GSL_func;
-    distExponentParams params = {ir,rGrid};
-    GSL_func.function = &(distExponentIntegral);
-    GSL_func.params = &params;
-    real_t abserr;
-    real_t epsabs = 0, epsrel = 3e-3, lim = gsl_ad_w->limit;
-    #define F(xi1,xi2,pitchDist) gsl_integration_qags(&GSL_func, xi1,xi2,epsabs,epsrel,lim,gsl_ad_w, &pitchDist, &abserr)
-    //////////////////////////////    
-
-    real_t dist1 = 0;
-    real_t dist2 = 0;
-
-    real_t thresholdToNeglectTrappedContribution = 1e-6;
-    if ( (xi0>xiT) || (xiT<thresholdToNeglectTrappedContribution) )
-        F(xi0,1.0,dist1);
-    else if ( (-xiT <= xi0) && (xi0 <= xiT) )
-        F(xiT,1.0,dist1);
-    else{ // (xi0 < -xiT)
-        F(xi0,-xiT,dist1);
-        F(xiT,1.0,dist2);
-    }
-    
-    #undef F
-    
-    return exp(-A*(dist1+dist2));
-}
-
-real_t EffectiveCriticalField::evaluatePitchDistribution(len_t ir, real_t xi0, real_t p, real_t Eterm, CollisionQuantity::collqty_settings *inSettings, gsl_integration_workspace *gsl_ad_w){
-    if(Eceff_mode == OptionConstants::COLLQTY_ECEFF_MODE_SIMPLE)
-        return evaluateApproximatePitchDistribution(ir,xi0,p,Eterm,inSettings);
-    else
-        return evaluateAnalyticPitchDistribution(ir,xi0,p,Eterm,inSettings,gsl_ad_w);
-}
 
 /*
 The function takes a xi0 and a lambda expression Func (and other needed helper parameters) and 
@@ -232,13 +171,12 @@ real_t UPartialContribution(real_t xi0, void *par){
     FVM::fluxGridType fluxGridType = params->fgType;
     gsl_integration_workspace *gsl_ad_w = params->gsl_ad_w;
     real_t E = params->Eterm;
-    EffectiveCriticalField *ecEff = params-> ecEff;
+    AnalyticDistribution *analyticDist = params-> analyticDist;
     std::function<real_t(real_t,real_t,real_t,real_t)> BAFunc = [xi0,params](real_t xiOverXi0,real_t BOverBmin,real_t /*ROverR0*/,real_t /*NablaR2*/){return params->Func(xi0,BOverBmin,xiOverXi0);};
     
     return rGrid->EvaluatePXiBounceIntegralAtP(ir,p,xi0,fluxGridType,BAFunc)
-        * ecEff->evaluatePitchDistribution(ir,xi0,p,E,collSettingsForEc, gsl_ad_w);
+        * analyticDist->evaluatePitchDistribution(ir,xi0,p,E,collSettingsForEc, gsl_ad_w);
 }
-
 
 /**
  * Public method used mainly for benchmarking: evaluates the pitch-averaged friction function -U 
@@ -250,8 +188,9 @@ real_t EffectiveCriticalField::testEvalU(len_t ir, real_t p, real_t Eterm, Colli
     const gsl_min_fminimizer_type *fmin_type = gsl_min_fminimizer_brent;
     gsl_min_fminimizer *fmin = gsl_min_fminimizer_alloc(fmin_type);
 
+    AnalyticDistribution analyticDist = AnalyticDistribution(rGrid, nuD, Eceff_mode);
     struct UContributionParams params = {rGrid, nuS,nuD, ir, p, FVM::FLUXGRIDTYPE_DISTRIBUTION, Eterm, Func, gsl_ad_w,
-                    fmin, p_ex_lo, p_ex_up,inSettings,QAG_KEY,this}; // @@Linnea move this to Eceff!
+                    fmin, p_ex_lo, p_ex_up,inSettings,QAG_KEY,&analyticDist}; 
     return UAtPFunc(p,&params);
 }
 
@@ -318,7 +257,6 @@ real_t EffectiveCriticalField::UAtPFunc(real_t p, void *par){
 
     real_t NuSContrib = -p*nuS->evaluateAtP(ir,p,collSettingsForEc) * UnityContrib;
 
-
     // Evaluates the contribution from synchrotron term A^p coefficient
     std::function<real_t(real_t,real_t,real_t)> FuncSynchrotron = 
             [](real_t xi0, real_t BOverBmin, real_t){return (1-xi0*xi0)*BOverBmin*BOverBmin*BOverBmin;};
@@ -341,35 +279,4 @@ real_t EffectiveCriticalField::UAtPFunc(real_t p, void *par){
 
     return -(EContrib + NuSContrib + SynchContrib) / UnityContrib;
 
-}
-
-/**
- * Same as evaluteAnalyticPitchDistribution, but approximating
- * xi0/<xi> = 1 for passing and 0 for trapped (thus avoiding the 
- * need for the numerical integration).
- */
-real_t EffectiveCriticalField::evaluateApproximatePitchDistribution(len_t ir, real_t xi0, real_t p, real_t Eterm, CollisionQuantity::collqty_settings *inSettings){
-    const real_t Bmin = rGrid->GetBmin(ir);
-    const real_t Bmax = rGrid->GetBmax(ir);
-    const real_t B2avgOverBmin2 = rGrid->GetFSA_B2(ir);
-    real_t xiT = sqrt(1-Bmin/Bmax);
-    real_t E = Constants::ec * Eterm / (Constants::me * Constants::c) * sqrt(B2avgOverBmin2); 
-
-    real_t pNuD = p*nuD->evaluateAtP(ir,p,inSettings);    
-    real_t A = 2*E/pNuD;
-
-    real_t dist1 = 0;
-    real_t dist2 = 0;
-
-    real_t thresholdToNeglectTrappedContribution = 1e-6;
-    if ( (xi0>xiT) || (xiT<thresholdToNeglectTrappedContribution) )
-        dist1 = 1-xi0;
-    else if ( (-xiT <= xi0) && (xi0 <= xiT) )
-        dist1 = 1-xiT;
-    else{ // (xi0 < -xiT)
-        dist1 = 1-xiT;
-        dist2 = -xiT - xi0;
-    }
-        
-    return exp(-A*(dist1+dist2));
 }
